+++ carla_ros_bridge/test/ros_bridge_client.py
@@ -134,6 +134,38 @@ class TestClock(unittest.TestCase):
         self.assertEqual(msg.width, 800)
         self.assertEqual(msg.encoding, "bgra8")
 
+    def test_dvs_camera_info(self):
+        """
+        Tests dvs camera info
+        """
+        rospy.init_node('test_node', anonymous=True)
+        msg = rospy.wait_for_message(
+            "/carla/ego_vehicle/camera/dvs/front/camera_info", CameraInfo, timeout=TIMEOUT)
+        self.assertEqual(msg.header.frame_id, "ego_vehicle/camera/dvs/front")
+        self.assertEqual(msg.height, 600)
+        self.assertEqual(msg.width, 800)
+
+    def test_dvs_camera_image(self):
+        """
+        Tests dvs camera images
+        """
+        rospy.init_node('test_node', anonymous=True)
+        msg = rospy.wait_for_message(
+            "/carla/ego_vehicle/camera/dvs/front/image_events", Image, timeout=TIMEOUT)
+        self.assertEqual(msg.header.frame_id, "ego_vehicle/camera/dvs/front")
+        self.assertEqual(msg.height, 600)
+        self.assertEqual(msg.width, 800)
+        self.assertEqual(msg.encoding, "bgr8")
+
+    def test_dvs_camera_events(self):
+        """
+        Tests dvs camera events
+        """
+        rospy.init_node('test_node', anonymous=True)
+        msg = rospy.wait_for_message(
+            "/carla/ego_vehicle/camera/dvs/front/events", PointCloud2, timeout=TIMEOUT)
+        self.assertEqual(msg.header.frame_id, "ego_vehicle/camera/dvs/front")
+
     def test_lidar(self):
         """
         Tests Lidar sensor node
@@ -142,6 +174,15 @@ class TestClock(unittest.TestCase):
         msg = rospy.wait_for_message(
             "/carla/ego_vehicle/lidar/lidar1/point_cloud", PointCloud2, timeout=TIMEOUT)
         self.assertEqual(msg.header.frame_id, "ego_vehicle/lidar/lidar1")
+
+    def test_semantic_lidar(self):
+        """
+        Tests semantic_lidar sensor node
+        """
+        rospy.init_node('test_node', anonymous=True)
+        msg = rospy.wait_for_message(
+            "/carla/ego_vehicle/semantic_lidar/lidar1/point_cloud", PointCloud2, timeout=TIMEOUT)
+        self.assertEqual(msg.header.frame_id, "ego_vehicle/semantic_lidar/lidar1")
 
     def test_radar(self):
         """